+++ CHANGELOG.md
@@ -7,9 +7,7 @@
 - New error types added.
 - Stubble options can be passed to `Stubble()` by the second parameter:
 
-Available options:
-    - `ignoreUnregisteredHelperErrors` - ignore errors of 'Unregistered helpers'
-    - `ignoreTagCaseSensetive` - ignore case of open and close tags 
+Available options: - `ignoreUnregisteredHelperErrors` - ignore errors of 'Unregistered helpers' - `ignoreTagCaseSensetive` - ignore case of open and close tags
 
 ## 1.0.2
 
@@ -22,3 +20,7 @@ Now with null-safety
 ## v 1.1.0
 
 - nested same named blocks available now
+
+## v 1.1.1
+
+- Empty or null templates are not caused an exception. Compiling for such templates will return an empty string